# scripts/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminOrderItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminOrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminOrderItemListResponse",
        "type": "object"
      },
      "AdminOrderItemResponse": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "fulfillmentStatus": {
            "enum": [
              "paid",
              "preparing",
              "shipped",
              "delivered"
            ],
            "title": "Fulfillmentstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lineTotalCredits": {
            "title": "Linetotalcredits",
            "type": "integer"
          },
          "orderId": {
            "title": "Orderid",
            "type": "string"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productTitle": {
            "title": "Producttitle",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "sellerType": {
            "enum": [
              "platform",
              "merchant"
            ],
            "title": "Sellertype",
            "type": "string"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          },
          "unitPriceCredits": {
            "title": "Unitpricecredits",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderId",
          "productId",
          "productTitle",
          "qty",
          "unitPriceCredits",
          "lineTotalCredits",
          "fulfillmentStatus",
          "shopName",
          "sellerType"
        ],
        "title": "AdminOrderItemResponse",
        "type": "object"
      },
...
      "CheckoutResponse": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderResponse"
          }
        },
        "required": [
          "order"
        ],
        "title": "CheckoutResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderItemResponse": {
        "properties": {
          "fulfillmentStatus": {
            "enum": [
              "paid",
              "preparing",
              "shipped",
              "delivered"
            ],
            "title": "Fulfillmentstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lineTotalCredits": {
            "title": "Linetotalcredits",
            "type": "integer"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productTitle": {
            "title": "Producttitle",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "sellerId": {
            "title": "Sellerid",
            "type": "string"
          },
          "sellerType": {
            "enum": [
              "platform",
              "merchant"
            ],
            "title": "Sellertype",
            "type": "string"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          },
          "unitPriceCredits": {
            "title": "Unitpricecredits",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "productId",
          "productTitle",
          "sellerId",
          "shopName",
          "sellerType",
          "qty",
          "unitPriceCredits",
          "lineTotalCredits",
          "fulfillmentStatus"
        ],
        "title": "OrderItemResponse",
        "type": "object"
      },
      "OrderListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "OrderListResponse",
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "paid",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "totalCredits": {
            "title": "Totalcredits",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "totalCredits",
          "items"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
      "ProductListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ProductListResponse",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "priceCredits": {
            "title": "Pricecredits",
            "type": "integer"
          },
          "seller": {
            "$ref": "#/components/schemas/SellerSummary"
          },
          "status": {
            "default": "published",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "stock": {
            "title": "Stock",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "priceCredits",
          "stock",
          "category",
          "seller"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
...
      "SellerOrderItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SellerOrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "SellerOrderItemListResponse",
        "type": "object"
      },
      "SellerOrderItemResponse": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "fulfillmentStatus": {
            "enum": [
              "paid",
              "preparing",
              "shipped",
              "delivered"
            ],
            "title": "Fulfillmentstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lineTotalCredits": {
            "title": "Linetotalcredits",
            "type": "integer"
          },
          "orderId": {
            "title": "Orderid",
            "type": "string"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productTitle": {
            "title": "Producttitle",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "unitPriceCredits": {
            "title": "Unitpricecredits",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderId",
          "productId",
          "productTitle",
          "qty",
          "unitPriceCredits",
          "lineTotalCredits",
          "fulfillmentStatus"
        ],
        "title": "SellerOrderItemResponse",
        "type": "object"
      },
...
      "SellerProductCreateRequest": {
        "properties": {
          "category": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "imageUrl": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "priceCredits": {
            "exclusiveMinimum": 0.0,
            "title": "Pricecredits",
            "type": "integer"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "stock": {
            "minimum": 0.0,
            "title": "Stock",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "priceCredits",
          "stock",
          "category"
        ],
        "title": "SellerProductCreateRequest",
        "type": "object"
      },
      "SellerProductUpdateRequest": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "imageUrl": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "priceCredits": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricecredits"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "published",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "stock": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "SellerProductUpdateRequest",
        "type": "object"
      },
...
      "SellerSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sellerType": {
            "enum": [
              "platform",
              "merchant"
            ],
            "title": "Sellertype",
            "type": "string"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          }
        },
        "required": [
          "id",
          "shopName",
          "sellerType"
        ],
        "title": "SellerSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Shopping mall backend skeleton",
    "title": "Shopping Mall API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/orders": {
      "get": {
        "operationId": "list_admin_orders_admin_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderItemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Admin Orders",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/me/orders": {
      "get": {
        "operationId": "read_orders_me_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Orders",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "create_order_me_orders_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/me/orders/{order_id}": {
      "get": {
        "operationId": "read_order_me_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Order",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/products": {
      "get": {
        "operationId": "get_products_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Products",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{product_id}": {
      "get": {
        "operationId": "get_product_by_id_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product By Id",
        "tags": [
          "products"
        ]
      }
    },
...
    "/seller/orders": {
      "get": {
        "operationId": "seller_list_orders_seller_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerOrderItemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seller List Orders",
        "tags": [
          "seller"
        ]
      }
    },
...
    "/seller/products": {
      "get": {
        "operationId": "seller_list_products_seller_products_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response Seller List Products Seller Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seller List Products",
        "tags": [
          "seller"
        ]
      },
      "post": {
        "operationId": "seller_create_product_seller_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seller Create Product",
        "tags": [
          "seller"
        ]
      }
    },
    "/seller/products/{product_id}": {
      "delete": {
        "operationId": "seller_delete_product_seller_products__product_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seller Delete Product",
        "tags": [
          "seller"
        ]
      },
      "patch": {
        "operationId": "seller_update_product_seller_products__product_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seller Update Product",
        "tags": [
          "seller"
        ]
      }
    }
  }
}